HHH-8404: Correctly render 'ListIndexExpression'.
A correctly rendered 'index' expression should be "index(foo)" where 'foo' is
the name (or alias) of the list table. Since 'ListAttributeJoin' passes 'this'
as 'origin' when creating a 'ListIndexExpression', "origin.getPathIdentifier()"
will already refer to this list table on the right-hand side of the join.
Appending the list attribute (which is relative to the table on the left-hand
side of the join) does not make sense.